vitessdriver: switch to go1.8
Issue #2681
We still can't get rid of the legacy 1.7 functions because of how
interfaces work. However, I assume that database/sql won't use them
if it finds the newer 'Context' functions in the driver.

Also, go1.9 has allowed us to implement support for custom types.
We still need to exploit this flexibility to add support for
unsigned and other exotic MySQL types.
